requirements.txt
google_api_python_client==2.154.0
moviepy==1.0.3
numpy==2.2.0
oauth2client==4.1.3
openai==1.57.3
openai_whisper==20240930
pillar_youtube_upload==0.3.0
soundfile==0.12.1
torch==2.2.2
yt_dlp==2024.12.13
